bq_reconciliation/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
pyyaml>=6.0
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
